Previous Colour
I have created a new span that allows the user to select the previous colour.

I have achieved this by creating a variable called "oldRGB" and another one called "newRGB", I assign the colours to these variables and use them depending on what the user clicks.

I had problems with when to assign the old and new colour to the variable but found out where to put the variable declaration so that the oldRGB isn't the same as the newRGB value.